Helps-2268 Add Deprecation message  (#573)
* finally!

* rename

* update according to comments
